churn: Make tools be plugins
- split tools into submodules
- use pluggy to register them as plugins
