+++ setup.cfg
@@ -43,6 +43,7 @@ install_requires =
     click-help-colors
     colorama  # for Windows
     importlib-metadata
+    pluggy
     pygments
     pyyaml
     rich >= 9.0
@@ -55,6 +56,11 @@ where = src
 [options.entry_points]
 console_scripts =
     mk=mk.__main__:cli
+mk_tools =
+    tox=mk.tools.tox:ToxTool
+    npm=mk.tools.npm:NpmTool
+    pre-commit=mk.tools.pre_commit:PreCommitTool
+    make=mk.tools.make:MakeTool
 
 [flake8]
 max-complexity = 22
