+++ src/mk/tools/__init__.py
@@ -1,6 +1,3 @@
-import os
-import subprocess
-from configparser import ConfigParser
 from typing import Any, List, Optional
 
 
@@ -69,64 +66,3 @@ class Tool(metaclass=ToolRegistry):
 
     def __rich__(self):
         return f"[magenta]{self.name}[/]"
-
-
-class PreCommitTool(Tool):
-    name = "pre-commit"
-
-    def run(self, action: Optional[str] = None):
-        subprocess.run(["pre-commit", "run", "-a"], check=True)
-
-    def is_present(self, path: str) -> bool:
-        if os.path.isfile(os.path.join(path, ".pre-commit-config.yaml")):
-            return True
-        return False
-
-    def actions(self) -> List[Action]:
-        return [Action(name="lint", tool=self)]
-
-
-class ToxTool(Tool):
-    name = "tox"
-
-    def is_present(self, path: str) -> bool:
-        if os.path.isfile(os.path.join(path, "tox.ini")):
-            return True
-        return False
-
-    def actions(self) -> List[Action]:
-        # -a is not supported by tox4!
-        actions: List[Action] = []
-        cp = ConfigParser()
-        tox_cfg = subprocess.check_output(["tox", "--showconfig"], universal_newlines=True)
-        cp.read_string(tox_cfg)
-        for section in cp.sections():
-            if section.startswith("testenv:"):
-                _, env_name = section.split(":")
-                # we ignore hidden envs like implicit .pkg:
-                if not env_name.startswith("."):
-                    actions.append(
-                        Action(
-                            name=env_name,
-                            tool=self,
-                            description=cp[section]["description"],
-                            args=[env_name],
-                        )
-                    )
-
-        return actions
-
-    def run(self, action: Optional[str] = None) -> None:
-        if not action:
-            cmd = ["tox"]
-        else:
-            cmd = ["tox", "-e", action]
-        subprocess.run(cmd, check=True)
-
-
-class MakeTool(Tool):
-    name = "make"
-
-
-class NpmTool(Tool):
-    name = "npm"
