+++ src/mk/tools/make.py
@@ -1,0 +1,5 @@
+from mk.tools import Tool
+
+
+class MakeTool(Tool):
+    name = "make"
