+++ src/mk/tools/npm.py
@@ -1,0 +1,5 @@
+from mk.tools import Tool
+
+
+class NpmTool(Tool):
+    name = "npm"
